requirements.txt
streamlit==1.49.1
streamlit-option-menu==0.4.0
requests==2.32.3

fastapi==0.115.5
uvicorn==0.32.1
scikit-learn==1.7.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
alembic==1.14.1
aiofiles==24.1.0
